+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Change Log
 
-### Ver 0.2.1(Next release, Not released)
+### Ver 0.2.2(Next release, Not released)
+
+### Ver 0.2.1
 
   * #5 [Enable to add badge / favicon like chrome](https://github.com/georgeOsdDev/react-draggable-tab/issues/5)
   * #6 [Add style on parent div](https://github.com/georgeOsdDev/react-draggable-tab/issues/6)